Renamed folders after the subtree split from squeezeslave
In order to clean up the repository and reduce its size to just the bare minimum (source code to squeezeslave and
libslimproto / libslimaudio itself), a git-subtree split was performed, placing the contents of the src directory in
a repo by itself.  This should allow us to track upstream while keeping our forked version intact.